Finished separating solution into different Projects

+++ Physics/src/Physics.h
@@ -10,7 +10,7 @@ namespace lge
 	vec2 getPosition(PhysicsBody* body);
 	void setPosition(PhysicsBody* body, vec2 position);
 	void applyGravity(PhysicsBody* body, double deltaTime, double timeSteps);
-	bool updatePhysicsBody(PhysicsBody* body, float deltaTime, float timeSteps);
+	bool updatePhysicsBody(PhysicsBody* body, double deltaTime, double timeSteps);
 	void constrainPhysicsBody(PhysicsBody* body, vec2 min, vec2 max);
 	void bounceInArea(PhysicsBody* body, vec2 min, vec2 max);
 }
